Move topology/mergepatch helper to an internal package
